Add new AddCommand class for generalizing add command in every level
