+++ data/admin/cs2113 1/topic 1.txt
@@ -1,0 +1,1 @@
+[Q] question | [A] answer
