+++ src/main/java/commands/AddChapterCommand.java
@@ -8,12 +8,12 @@ import access.Access;
 import storage.Storage;
 import ui.Ui;
 
-public class AddChapterCommand extends Command {
-    public static final String COMMAND_WORD = "addchapter";
+public class AddChapterCommand extends AddCommand {
     private final Chapter chapter;
+    public static final String CHAPTER_PARAMETERS = " CHAPTER_NAME";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Adds a chapter to the module. \n"
-            + "Parameters: CHAPTER_NAME\n"
+            + "Parameters:" + CHAPTER_PARAMETERS + "\n"
             + "Example: " + COMMAND_WORD + " Chapter 1\n";
 
     public AddChapterCommand(String chapterCode) {
